backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
pydantic==2.10.3
pydantic-settings==2.6.1
numpy==1.26.4
openai-whisper==20240930
deepface==0.0.93
tf-keras==2.18.0
openai==1.58.1
anthropic==0.40.0
opencv-python-headless==4.10.0.84
httpx==0.28.1
aiofiles==24.1.0
pytest==8.3.4
pytest-asyncio==0.24.0
